requirements.txt
hmmlearn==0.3.2
numpy>=1.24
pandas>=2.0
pybit>=5.14.0
httpx>=0.27
pyyaml>=6.0
scipy>=1.12
scikit-learn>=1.3.0
python-dotenv>=1.0.0
lightgbm>=4.0.0
pyarrow>=14.0.0
